test: 로또 번호 6개 테스트 추가
